refactor helpers in interpreter/craftassist
