+++ droidlet/interpreter/craftassist/dummy_interpreter.py
@@ -6,7 +6,7 @@ from typing import Tuple, Dict, Any, Optional
 from droidlet.dialog.dialogue_objects import DialogueObject
 from ..interpreter import ReferenceObjectInterpreter, FilterInterpreter, interpret_reference_object
 from ..interpret_conditions import ConditionInterpreter
-from .attribute_helper import MCAttributeInterpreter
+from .interpret_attributes import MCAttributeInterpreter
 
 
 class DummyInterpreter(DialogueObject):
