+++ droidlet/interpreter/craftassist/interpret_attributes.py
